+++ src/QuadTree.h
@@ -13,7 +13,7 @@ public:
 	void insert(Point p);
 	void list(std::vector<PointList> &v);
 	int countRegion(Point p, int d);
-	int aggregateRegion(Point p, int d);
+	long long aggregateRegion(Point p, int d);
 
 private:
 	Node* root;
